EXCEPTION
Write a Calculator class with a single method: int power(int,int). The power method takes two integers,  N and , P as parameters and returns the integer result of N^P . If either  or  is negative, then the method must throw an exception with the message: n and p should be non-negative.